logout button removed from KingdomOverviewFragment, and put into OptionsMenu logout
