+++ app/src/main/res/layout/kingdom_overview_layout.xml
@@ -12,12 +12,4 @@
         android:layout_marginTop="151dp"
         android:id="@+id/kingdomOverview" />
 
-    <Button
-        android:text="Logout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="logout"
-        android:id="@+id/logoutButton"/>
 </RelativeLayout>
